tests: move to separate folder
